packaging: relocate sources under the fail2ban_exporter package
