+++ fail2ban_exporter/__init__.py
@@ -1,0 +1,1 @@
+"""A Prometheus exporter for fail2ban jail data"""
